backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
alembic==1.13.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
eval_type_backport==0.2.2
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1
